Struct static_rc::rcref::StaticRcRef

pub struct StaticRcRef<'a, T: ?Sized, const NUM: usize, const DEN: usize> { /* fields omitted */ }

A compile-time reference-counted pointer.

The inherent methods of StaticRc are all associated functions to avoid conflicts with the the methods of the inner type T which are brought into scope by the Deref implementation.

The parameters NUM and DEN DENote the ratio (NUM / DEN) of ownership of the pointer:

The ratio is always in the (0, 1] interval, that is: NUM > 0 and NUM <= DEN.

When the ratio is equal to 1, that is when NUM == DEN, then the instance has full ownership of the pointee and extra capabilities are unlocked.

Implementations

impl<'a, T: ?Sized, const N: usize> StaticRcRef<'a, T, N, N>

pub fn new(value: &'a mut T) -> Self where
(): Sized,

Constructs a new StaticRcRef<'a, T, N, N>.

Example

use static_rc::StaticRcRef;

type Full<'a> = StaticRcRef<'a, i32, 1, 1>;

let mut value = 42;
let rc = Full::new(&mut value);
assert_eq!(42, *rc);

pub fn pin(value: &'a mut T) -> Pin<Self> where
(): Sized,

Constructs a new Pin<StaticRcRef<'a, T, N, N>>.

Example

use static_rc::StaticRcRef;

type Full<'a> = StaticRcRef<'a, i32, 1, 1>;

let mut value = 42;
let rc = Full::pin(&mut value);
assert_eq!(42, *rc);

pub fn into_inner(this: Self) -> &'a mut T

Returns the inner value.

Example

use static_rc::StaticRcRef;

type Full<'a> = StaticRcRef<'a, i32, 1, 1>;

let mut value = 42;
let rc = Full::new(&mut value);
let inner: &mut i32 = Full::into_inner(rc);
assert_eq!(42, *inner);

pub fn get_mut(this: &mut Self) -> &mut T

Returns a mutable reference into the given StaticRcRef.

Example

use static_rc::StaticRcRef;

type Full<'a> = StaticRcRef<'a, i32, 1, 1>;

let mut value = 42;
let mut rc = Full::new(&mut value);
assert_eq!(42, *Full::get_mut(&mut rc));

impl<'a, T: ?Sized, const NUM: usize, const DEN: usize> StaticRcRef<'a, T, NUM, DEN>

pub fn into_raw(this: Self) -> NonNull<T>

Consumes the StaticRcRef, returning the wrapped pointer.

Example

use static_rc::StaticRcRef;

type Full<'a> = StaticRcRef<'a, i32, 1, 1>;

let mut value = 42;
let rc = Full::new(&mut value);
let pointer = Full::into_raw(rc);
assert_eq!(&mut value as *mut _, pointer.as_ptr());

pub fn as_ptr(this: &Self) -> NonNull<T>

Provides a raw pointer to the data.

StaticRcRef is not consumed or affected in any way, the pointer is valid as long as the original value is.

Example

use static_rc::StaticRcRef;

type Full<'a> = StaticRcRef<'a, i32, 1, 1>;

let mut value = 42;
let pointer = &mut value as *mut _;

let rc = Full::new(&mut value);
let other_pointer = Full::as_ptr(&rc);

assert_eq!(pointer, other_pointer.as_ptr());

pub fn get_ref(this: &Self) -> &T

Provides a reference to the data.

Example

use static_rc::StaticRcRef;

type Full<'a> = StaticRcRef<'a, i32, 1, 1>;

let mut value = 42;
let rc = Full::new(&mut value);

let r = Full::get_ref(&rc);

assert_eq!(42, *r);

pub unsafe fn from_raw(pointer: NonNull<T>) -> Self where
(): Sized,

Constructs a StaticRcRef<'a, T, NUM, DEN> from a raw pointer.

The raw pointer must have been previously returned by a call to StaticRcRef<'a, U, N, D>::into_raw:

If U is different from T, then specific restrictions on size and alignment apply. See mem::transmute for the restrictions applying to transmuting references.

If N / D is different from NUM / DEN, then specific restrictions apply. The user is responsible for ensuring proper management of the ratio of shares, and ultimately that the value is not dropped twice.

pub fn ptr_eq<const N: usize, const D: usize>(
this: &Self,
other: &StaticRcRef<'a, T, N, D>
) -> bool

Returns true if the two StaticRcRef point to the same allocation.

Example

use static_rc::StaticRcRef;

type Full<'a> = StaticRcRef<'a, i32, 2, 2>;

let mut value = 42;
let rc = Full::new(&mut value);
let (one, two) = Full::split::<1, 1>(rc);

assert!(StaticRcRef::ptr_eq(&one, &two));

pub fn adjust<const N: usize, const D: usize>(
this: Self
) -> StaticRcRef<'a, T, N, D> where
(): Sized,
((), ()): Sized,

Adjusts the NUMerator and DENumerator of the ratio of the instance, preserving the ratio.

Panics

If the compile-time-ratio feature is not used, and the ratio is not preserved; that is N / D <> NUM / DEN.

Example

use static_rc::StaticRcRef;

type Full<'a> = StaticRcRef<'a, i32, 2, 2>;

let mut value = 42;
let rc = Full::new(&mut value);
let rc = Full::adjust::<1, 1>(rc);

assert_eq!(42, *rc);

pub fn split<const A: usize, const B: usize>(
this: Self
) -> (StaticRcRef<'a, T, A, DEN>, StaticRcRef<'a, T, B, DEN>) where
(): Sized,
(): Sized,
((), ()): Sized,

Splits the current instance into two instances with the specified NUMerators.

Panics

If the compile-time-ratio feature is not used, and the ratio is not preserved; that is A + B <> NUM.

Example

use static_rc::StaticRcRef;

type Full<'a> = StaticRcRef<'a, i32, 2, 2>;

let mut value = 42;
let rc = Full::new(&mut value);
let (one, two) = Full::split::<1, 1>(rc);

assert_eq!(42, *one);
assert_eq!(42, *two);

pub fn split_array<const N: usize, const DIM: usize>(
this: Self
) -> [StaticRcRef<'a, T, N, DEN>; DIM] where
T: 'a,
((), ()): Sized,
(): Sized,

Splits the current instance into DIM instances with the specified Numerators.

Panics

If the compile-time-ratio feature is not used, and the ratio is not preserved; that is N * DIM <> NUM.

Example

use static_rc::StaticRcRef;

type Full<'a> = StaticRcRef<'a, i32, 2, 2>;

let mut value = 42;
let rc = Full::new(&mut value);
let array = Full::split_array::<1, 2>(rc);

assert_eq!(42, *array[0]);
assert_eq!(42, *array[1]);

pub fn join<const A: usize, const B: usize>(
left: StaticRcRef<'a, T, A, DEN>,
right: StaticRcRef<'a, T, B, DEN>
) -> Self

Joins two instances into a single instance.

Panics

If the two instances do no point to the same allocation, as determined by StaticRcRef::ptr_eq.

If the compile-time-ratio feature is not used and the ratio is not preserved; that is A + B <> NUM.

Example

use static_rc::StaticRcRef;

type Full<'a> = StaticRcRef<'a, i32, 2, 2>;

let mut value = 42;
let rc = Full::new(&mut value);
let (one, two) = Full::split::<1, 1>(rc);
let rc = Full::join(one, two);

assert_eq!(42, *rc);

pub unsafe fn join_unchecked<const A: usize, const B: usize>(
left: StaticRcRef<'a, T, A, DEN>,
_right: StaticRcRef<'a, T, B, DEN>
) -> Self

Joins two instances into a single instance without checking whether they point to the same allocation.

Unless compile-time-ratio is activated, the ratios are checked nevertheless.

Safety

The caller must guarantee that those instances point to the same allocation.

Panics

If the compile-time-ratio feature is not used and the ratio is not preserved; that is A + B <> NUM.

In debug, if the two instances do no point to the same allocation, as determined by StaticRcRef::ptr_eq.

Example

use static_rc::StaticRcRef;

type Full<'a> = StaticRcRef<'a, i32, 2, 2>;

let mut value = 42;
let rc = Full::new(&mut value);
let (one, two) = Full::split::<1, 1>(rc);
let rc = unsafe { Full::join_unchecked(one, two) };

assert_eq!(42, *rc);

pub fn join_array<const N: usize, const DIM: usize>(
array: [StaticRcRef<'a, T, N, DEN>; DIM]
) -> Self

Joins DIM instances into a single instance.

Panics

If all instances do not point to the same allocation, as determined by StaticRcRef::ptr_eq.

If the compile-time-ratio feature is not used and the ratio is not preserved; that is N * DIM <> NUM.

Example

use static_rc::StaticRcRef;

type Full<'a> = StaticRcRef<'a, i32, 2, 2>;

let mut value = 42;
let rc = Full::new(&mut value);
let array = Full::split_array::<1, 2>(rc);
let rc = Full::join_array(array);

assert_eq!(42, *rc);

pub unsafe fn join_array_unchecked<const N: usize, const DIM: usize>(
array: [StaticRcRef<'a, T, N, DEN>; DIM]
) -> Self

Joins DIM instances into a single instance.

Panics

If the compile-time-ratio feature is not used and the ratio is not preserved; that is N * DIM <> NUM.

In debug, if all instances do not point to the same allocation, as determined by StaticRcRef::ptr_eq.

Example

use static_rc::StaticRcRef;

type Full<'a> = StaticRcRef<'a, i32, 2, 2>;

let mut value = 42;
let rc = Full::new(&mut value);
let array = Full::split_array::<1, 2>(rc);
let rc = unsafe { Full::join_array_unchecked(array) };

assert_eq!(42, *rc);

impl<'a, const NUM: usize, const DEN: usize> StaticRcRef<'a, dyn Any, NUM, DEN>

pub fn downcast<T: Any>(self) -> Result<StaticRcRef<'a, T, NUM, DEN>, Self>

Attempts to downcast Self to a concrete type.